requirements.txt
paho-mqtt
sqlite3
